+++ R/server.R
@@ -8,44 +8,69 @@ server <- function(input, output) {
         
         click <- input$Next
         
-        #skip first page?
-        if(!length(.MCE$shinyGUI$firstpage)) click <- click + 1L
-        
-        #first page, ask for demographics, etc
-        if(click == 0L){
-            return(.MCE$shinyGUI$firstpage)
+        if(length(.MCE$shinyGUI$password)){
+            if(click == 0L){
+                if(nrow(.MCE$shinyGUI$password) > 1L)
+                    return(list(textInput('UsErNaMe', label = "Login Name:"),
+                                passwordInput('PaSsWoRd', 'Password:')))
+                else return(passwordInput('PaSsWoRd', 'Password:'))
+            } else if(click == 1L){
+                .MCE$verified <- verifyPassword(input, .MCE$shinyGUI$password)
+            }
+            click <- click - 1L
         }
         
-        #skip demographics page?
-        if(!length(.MCE$shinyGUI$demographics)) click <- click + 1L
+        if(!.MCE$verified)
+            return(h3('Incorrect Login Name/Password. Please restart the application and try again.'))
         
-        if(click == 1L){
-            return(.MCE$shinyGUI$demographics)
-        }
+        if(.MCE$resume_file && click < 1L){
+            return(list(h5("Click the action button to continue with your session.")))
+        } else {
+            #skip first page? Demographics, etc
+            if(!length(.MCE$shinyGUI$firstpage)) click <- click + 1L
+            if(click == 0L)
+                return(.MCE$shinyGUI$firstpage)
+            
+            #skip demographics page?
+            if(!length(.MCE$shinyGUI$demographics)) click <- click + 1L
+            if(click == 1L)
+                return(.MCE$shinyGUI$demographics)
+            
+            #store demographic results
+            if(click == 2L){
+                tmp <- list()
+                for(tag in .MCE$shinyGUI$demographic_inputIDs)
+                    tmp[[length(tmp) + 1L]] <- input[[tag]]
+                names(tmp) <- .MCE$shinyGUI$demographic_inputIDs
+                .MCE$person$field("demographics", as.data.frame(tmp))
+                if(!is.null(.MCE$last_demographics))
+                    .MCE$person$demographics <- .MCE$last_demographics
+                if(.MCE$shinyGUI$temp_file != '')
+                    saveRDS(.MCE$person, .MCE$shinyGUI$temp_file)
+            }
+            
+            if(.MCE$shinyGUI$begin_message == "") click <- click + 1L
+            if(click == 2L)
+                return(list(h5(.MCE$shinyGUI$begin_message)))
+        } #end normal start
         
-        #store demographic results
-        if(click == 2L){
-            tmp <- list()
-            for(tag in .MCE$shinyGUI$demographic_inputIDs)
-                tmp[[length(tmp) + 1L]] <- input[[tag]]
-            names(tmp) <- .MCE$shinyGUI$demographic_inputIDs
-            .MCE$person$field("demographics", as.data.frame(tmp))
-            if(!is.null(.MCE$last_demographics))
-                .MCE$person$demographics <- .MCE$last_demographics
-            if(.MCE$shinyGUI$temp_file != '')
-                saveRDS(.MCE$person, .MCE$shinyGUI$temp_file)
-            return(list(h5(.MCE$shinyGUI$begin_message)))
-        }
-        
-        if(click == 3L) .MCE$start_time <- proc.time()[3L]
+        if(is.null(.MCE$start_time))
+            .MCE$start_time <- proc.time()[3L]
         
         if(.MCE$resume_file){
             .MCE$resume_file <- FALSE
             item <- max(which(!is.na(.MCE$person$items_answered)))
-            return(list(.MCE$shinyGUI$df$Question[[item]], .MCE$shinyGUI$questions[[item]]))
+            stemOutput <- stemContent(item)
+            return(list(stemOutput,.MCE$shinyGUI$df$Question[[item]], 
+                        .MCE$shinyGUI$questions[[item]]))
         }
         
         itemclick <- sum(!is.na(.MCE$person$items_answered))
+        
+        if(FALSE){
+            cat('\nclick = ', click)
+            cat('\titemclick = ', itemclick)
+        }
         
         # run survey
         if(click > 2L && !.MCE$design@stop_now){
@@ -53,6 +78,8 @@ server <- function(input, output) {
                 pick <- .MCE$person$items_answered[itemclick]
                 name <- .MCE$test@itemnames[pick]
                 ip <- unname(input[[name]])
+                if(.MCE$shinyGUI$df$Type[pick] == 'select' && .MCE$shinyGUI$forced_choice && ip == "")
+                    ip <- NULL
                 if(is.null(ip)) ip <- input[[paste0(.MCE$invalid_count, '.TeMpInTeRnAl',name)]]
                 if(!is.null(ip)){
                     ip <- as.character(ip)
@@ -69,8 +96,8 @@ server <- function(input, output) {
                             .MCE$person$responses[pick] <- as.integer(sum(ip %in% .MCE$test@item_answers[[pick]]))
                         else .MCE$person$responses[pick] <- as.integer(ip %in% .MCE$test@item_answers[[pick]])
                     }
-                    .MCE$person$item_time[pick] <- proc.time()[3L] - .MCE$start_time - 
-                        sum(.MCE$person$item_time)
+                    .MCE$person$item_time[pick] <- proc.time()[3L] - .MCE$start_time
+                    .MCE$start_time <- NULL
                     
                     #update Thetas
                     .MCE$person$Update.thetas(.MCE$design, .MCE$test)
@@ -83,10 +110,12 @@ server <- function(input, output) {
                         .MCE$invalid_count <- .MCE$invalid_count + 1L
                         tmp <- lapply(.MCE$shinyGUI$df, function(x, pick) x[pick], pick=pick)
                         tmp <- buildShinyElements(tmp, paste0(.MCE$invalid_count, '.TeMpInTeRnAl', name))
-                        return(list(.MCE$shinyGUI$df$Question[[pick]], tmp$questions))
+                        stemOutput <- stemContent(pick)
+                        return(list(stemOutput, .MCE$shinyGUI$df$Question[[pick]], 
+                                    tmp$questions))
                     } else {
-                        .MCE$person$item_time[pick] <- proc.time()[3L] - .MCE$start_time - 
-                            sum(.MCE$person$item_time)
+                        .MCE$person$item_time[pick] <- proc.time()[3L] - .MCE$start_time
+                        .MCE$start_time <- NULL
                         #update Thetas (same as above)
                         .MCE$person$Update.thetas(.MCE$design, .MCE$test)
                         if(.MCE$shinyGUI$temp_file != '')
@@ -105,10 +134,15 @@ server <- function(input, output) {
                 if(is.na(item)){
                     .MCE$design@stop_now <- TRUE
                 } else {
+                    if(is.null(.MCE$start_time))
+                        .MCE$start_time <- proc.time()[3L]
                     .MCE$person$items_answered[itemclick+1L] <- item
                     if(.MCE$shinyGUI$temp_file != '')
                         saveRDS(.MCE$person, .MCE$shinyGUI$temp_file)
-                    return(list(.MCE$shinyGUI$df$Question[[item]], .MCE$shinyGUI$questions[[item]]))
+                    stemOutput <- stemContent(item)
+                    return(list(stemOutput, 
+                                .MCE$shinyGUI$df$Question[[item]], 
+                                .MCE$shinyGUI$questions[[item]]))
                 }
             }
         }
@@ -129,35 +163,5 @@ server <- function(input, output) {
             return(NULL)
         }
         
-    }) 
-        
-    output$item_stem_html <- renderUI({
-        
-        click <- input$Next - .MCE$shift_back
-        if(click > 0L)
-            if(!length(.MCE$shinyGUI$demographics)) click <- click + 1L
-        
-        if(!.MCE$STOP){
-            if(click > 2L && (click-2L) < .MCE$test@length){
-                pick <- force(.MCE$person$items_answered[[click-2L]])
-                file <- .MCE$shinyGUI$stem_locations[pick]
-                empty <- is.na(file)
-                if(!empty){
-                    if(grepl('\\.[mM][dD]$', file)){
-                        suppressWarnings(markdown::markdownToHTML(file=file, output=.MCE$outfile2, 
-                                                 fragment.only = TRUE))
-                        contents <- readLines(.MCE$outfile2, warn = FALSE)
-                        return(HTML(contents))
-                    } else if(grepl('\\.[hH][tT][mM][lL]$', file)){
-                        contents <- readLines(file, warn = FALSE)
-                        return(HTML(contents))
-                    } else empty <- TRUE
-                }
-            } else empty <- TRUE
-        } else empty <- TRUE
-        
-        return(' ')
-        
     })
-    
 }